+++ app/build.gradle
@@ -57,9 +57,6 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
 
-    // koala nlp
-    implementation "kr.bydelta:koalanlp-hnn:2.1.4:assembly"
-
     // room
     implementation "androidx.room:room-runtime:$roomVersion"
     kapt "androidx.room:room-compiler:$roomVersion"
