+++ nottokeyword/build.gradle
@@ -44,4 +44,7 @@ dependencies {
     implementation platform('com.google.firebase:firebase-bom:29.0.0')
     implementation 'com.google.firebase:firebase-database-ktx'
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-play-services:1.2.1"
+
+    // koala nlp
+    implementation "kr.bydelta:koalanlp-hnn:2.1.4:assembly"
 }
